+++ packages/plugin-table/web/src/toolbar/buttons/BorderButton.tsx
@@ -16,7 +16,6 @@ const BorderButton: FC<Props> = ({ dataHook, toolbarItem }) => {
   const modalService = useContext(ModalContext) || {};
 
   const [referenceElement, setReferenceElement] = useState<HTMLDivElement | null>(null);
-  const [isButtonActive, setIsButtonActive] = useState(false);
 
   const isMultipleSelection = true; // TODO: selectedCells.length > 1
 
@@ -41,13 +40,6 @@ const BorderButton: FC<Props> = ({ dataHook, toolbarItem }) => {
       : closeModal();
   };
 
-  modalService.onModalOpened(() => {
-    modalService.isModalOpen(TABLE_BUTTONS_MODALS_ID.BORDER) && setIsButtonActive(true);
-  });
-  modalService.onModalClosed(() => {
-    !modalService.isModalOpen(TABLE_BUTTONS_MODALS_ID.BORDER) && setIsButtonActive(false);
-  });
-
   return !isMultipleSelection ? (
     <ToggleButton
       Icon={BorderIcon}
@@ -58,7 +50,7 @@ const BorderButton: FC<Props> = ({ dataHook, toolbarItem }) => {
   ) : (
     <DropdownButton
       dataHook={dataHook}
-      active={isButtonActive}
+      active={modalService.isModalOpen(TABLE_BUTTONS_MODALS_ID.BORDER)}
       onClick={openBorderPanel}
       setRef={setReferenceElement}
       Icon={BorderIcon}